recvmsg01: kill child only if one was forked
If start_server() fails, then pid will end up with -1 and
cleanup() will do kill(-1, SIGKILL), which kills everything
this process can kill.

Change the test to kill pid (child) only when it successfully
creates one.

+++ testcases/kernel/syscalls/recvmsg/recvmsg01.c
@@ -281,7 +281,8 @@ void setup(void)
 
 void cleanup(void)
 {
-	(void)kill(pid, SIGKILL);	/* kill server */
+	if (pid > 0)
+		(void)kill(pid, SIGKILL);	/* kill server */
 	if (tmpsunpath[0] != '\0')
 		(void)unlink(tmpsunpath);
 	TEST_CLEANUP;
